EOF_TRASH_CLIENT/requirements.txt
# python --verison 3.9.2
RPi.GPIO==0.7.1

opencv-python==4.8.1.78

#  
PyAudio==0.2.14
